Add selecitonWatcher#getFreshRange
Use the RangeContainer in the dispatcher. Also remove all occurrences of 'instanceof' to distinguish between Cursor and Selection. Use the new #isCursor and #isSelection properties of RangeContainer, Selection or Cursor instances.

Also use the new #isAtTextEnd method of Cursor to detect if the cursor is at the visible end of its host. This ignores following whitespaces, so users don't have to guess if they are at the end of the host. Also with non-breaking spaces it is not always possible to move the cursor at the very end of its host, thus making it sometimes impossible to for example create a new block with an Enter keystroke.