Add linking functionality
Add content#forceWrap() to make it possible to wrap content with a tag without producing odd markup. The method uses nuke internally to remove any other nodes of the same tagName that affect the selection prior to wrap the selection. If the selection is not wrappable, all tags that affect the selection are removed.