Add formatting methods to selection
I implemented a basic toolbar in main.js to test the new methods on the selection. After performing a formatting command on the selection the resulting selection is not always as it should be. After removing the formatting the selection will not be restored correctly and the surround command has some bugs with selection restoring too. Toggling of bold and the like is not implemented, just issuing simple commands.