Add specs for the dispatcher
There are multiple ways to test the dispatcher. For now I choose to test the whole stack with an initialized Editable. An alternative would be to call the `keyboard.dispatchKeyEvent` manually with a mocked event object. I needed a few helpers to catch the events fired by the dispatcher and a contenteditable block is appendend to document.body for every test. I found that the module pattern we use a lot can sometimes make it difficult in tests to observe what is happening. So I think we should try to improve testability of our code going forward.