Add Selection#toggleSurround()
I added methods to content to check whether a range contains a string and to remove a character from a range. With these the toggleSurround() method on the selection becomes possible. To remove all characters from a textnode and not just the first one I needed to convert the character into a regex with the global flag turned on. For this I added a helper to the string util which escapes all special characters.