Add content#isExactSelection()
Added a method to check if a tag is selected exactly. Nothing less or more. This is a bit tricky since ranges that look exactly the same to the user can look quite differently internally. To make sure the method works consistently I added tests for a few different scenarios (but the tests cover not everything, mind you).

The intention behind this is to make it possible to expand the selection to the whole tag in cases where the tag is not selected exactly.