Move cursor after text node
Before save the cursor position I ensure its position is
after the latest text node and not at the end of the
container element. In this way, the range save/restore
marker is after a text node and not between HTML
elements. This ensure that all the cleaning logic
(i.e. normalizeTags()) can work without any change.